requirements.txt
Pillow==6.2.1
RPi.GPIO==0.7.0
spidev==3.4
